Prepare build pipeline for new version of Testcontainers .NET, closes #41.
