+++ .gitattributes
@@ -1,0 +1,6 @@
+# Autodetect text files
+* text=auto
+
+# Definitively text files 
+*.cs text
+*.cake text
