+++ .gitignore
@@ -1,3 +1,6 @@
+# macOS
+.DS_Store
+
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
 ##
@@ -340,3 +343,14 @@ ASALocalRun/
 
 # BeatPulse healthcheck temp database
 healthchecksdb
+
+# Backup folder for Package Reference Convert tool in Visual Studio 2017
+MigrationBackup/
+
+# Cakebuild
+tools/*
+!tools/packages.config
+
+# Project
+test-results/
+test-coverage/
