--- Directory.Build.props
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Project>
-  <!-- Adds the properties below to each project -->
-  <PropertyGroup Condition="'$(BuildingInsideVisualStudio)' != 'true'">
-    <!-- Treat warnings from StyleCopAnalyzers as errors and break build -->
-    <TreatWarningsAsErrors>false</TreatWarningsAsErrors> <!-- Set this to true as soon as all warnings are fixed. -->
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Debug'">
-    <DebugType>full</DebugType>
-    <DebugSymbols>true</DebugSymbols>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Release'">
-    <DebugType>pdbonly</DebugType>
-    <DebugSymbols>true</DebugSymbols>
-  </PropertyGroup>
-</Project>
