--- Directory.Build.targets
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Project>
-  <!-- Runs tasks for different targets and build configurations -->
-</Project>
