--- TestContainers.Tests/TestContainers.Tests.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks Condition="'$(OS)' == 'Windows_NT'">net46;netcoreapp2.0</TargetFrameworks>
-    <TargetFrameworks Condition="'$(OS)' != 'Windows_NT'">netcoreapp2.0</TargetFrameworks>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-    <GenerateDocumentationFile>false</GenerateDocumentationFile>
-    <DebugType>full</DebugType>
-  </PropertyGroup>
-  <ItemGroup>
-    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
-    <PackageReference Include="coverlet.msbuild" Version="2.0.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.7.2" />
-    <PackageReference Include="MySql.Data" Version="6.10.6" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
-    <PackageReference Include="xunit" Version="2.3.1" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\TestContainers\TestContainers.csproj" />
-  </ItemGroup>
-  <Target Name="CopyRedisConfig" AfterTargets="AfterBuild">
-    <Copy SourceFiles="../RedisConfigs/master-6379.conf" DestinationFolder="$(Outdir)" />
-  </Target>
-</Project>
