+++ TestContainers.sln
@@ -1,55 +1,38 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestContainers", "TestContainers\TestContainers.csproj", "{CEBE0763-146E-4A92-8027-7403F099AE9B}"
+VisualStudioVersion = 15.0.28307.645
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{BB91417F-1345-4FAD-9D56-7B11DF593084}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestContainers.Tests", "TestContainers.Tests\TestContainers.Tests.csproj", "{C4449BD8-4483-438D-B9F2-C548F361012D}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{5DC26A60-8775-4076-A94D-1BDC36B0A828}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "RedisConfigs", "RedisConfigs", "{4C100A84-CE31-41D0-9191-4FBE6FE1FBB3}"
-	ProjectSection(SolutionItems) = preProject
-		RedisConfigs\master-6379.conf = RedisConfigs\master-6379.conf
-	EndProjectSection
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestContainers", "src\TestContainers\TestContainers.csproj", "{4D05E9EF-78BF-48A6-9CF6-D4F6AEBE2C7A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestContainers.Tests", "test\TestContainers.Tests\TestContainers.Tests.csproj", "{9C954639-D343-4E02-A5C8-B3A466CE2F70}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Debug|x64.Build.0 = Debug|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Debug|x86.Build.0 = Debug|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Release|x64.ActiveCfg = Release|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Release|x64.Build.0 = Release|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Release|x86.ActiveCfg = Release|Any CPU
-		{CEBE0763-146E-4A92-8027-7403F099AE9B}.Release|x86.Build.0 = Release|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Debug|x64.Build.0 = Debug|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Debug|x86.Build.0 = Debug|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Release|x64.ActiveCfg = Release|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Release|x64.Build.0 = Release|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Release|x86.ActiveCfg = Release|Any CPU
-		{C4449BD8-4483-438D-B9F2-C548F361012D}.Release|x86.Build.0 = Release|Any CPU
+		{4D05E9EF-78BF-48A6-9CF6-D4F6AEBE2C7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4D05E9EF-78BF-48A6-9CF6-D4F6AEBE2C7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4D05E9EF-78BF-48A6-9CF6-D4F6AEBE2C7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4D05E9EF-78BF-48A6-9CF6-D4F6AEBE2C7A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C954639-D343-4E02-A5C8-B3A466CE2F70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C954639-D343-4E02-A5C8-B3A466CE2F70}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C954639-D343-4E02-A5C8-B3A466CE2F70}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C954639-D343-4E02-A5C8-B3A466CE2F70}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{4D05E9EF-78BF-48A6-9CF6-D4F6AEBE2C7A} = {BB91417F-1345-4FAD-9D56-7B11DF593084}
+		{9C954639-D343-4E02-A5C8-B3A466CE2F70} = {5DC26A60-8775-4076-A94D-1BDC36B0A828}
+	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F385B2E2-A1ED-41A3-A68F-662C381EE606}
+		SolutionGuid = {6D84DD44-76E1-4131-9E5E-075E384BFEE6}
 	EndGlobalSection
 EndGlobal
