--- TestContainers/TestContainers.csproj
@@ -1,32 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks Condition="'$(OS)' == 'Windows_NT'">net46;netstandard2.0</TargetFrameworks>
-    <TargetFrameworks Condition="'$(OS)' != 'Windows_NT'">netstandard2.0</TargetFrameworks>
-    <Version>0.0.2</Version>
-    <Company>testcontainers.org</Company>
-    <Authors>swissarmykirpan</Authors>
-    <Product>TestContainers</Product>
-    <PackageId>TestContainers</PackageId>
-    <AssemblyName>TestContainers</AssemblyName>
-    <ProjectUrl>https://testcontainers.org</ProjectUrl>
-    <PackageLicenseUrl>https://raw.githubusercontent.com/testcontainers/testcontainers-dotnet/master/LICENSE</PackageLicenseUrl>
-    <PackageIconUrl>https://avatars0.githubusercontent.com/u/13393021?s=200&amp;v=4</PackageIconUrl>
-    <Summary></Summary>
-    <ReleaseNotes></ReleaseNotes>
-    <Language></Language>
-    <PackageTags>testcontainers</PackageTags>
-    <Description />
-    <Copyright>Copyright © 2018</Copyright>
-    <RepositoryUrl>https://github.com/testcontainers/testcontainers-dotnet</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Docker.Dotnet" Version="3.125.2" />
-    <PackageReference Include="MySql.Data" Version="6.10.6" />
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-    <PackageReference Include="Npgsql" Version="4.0.2" />
-    <PackageReference Include="Polly" Version="6.0.1" />
-    <PackageReference Include="RabbitMQ.Client" Version="5.1.0" />
-    <PackageReference Include="StackExchange.Redis" Version="1.2.6" />
-  </ItemGroup>
-</Project>
