+++ build.cake
@@ -1,0 +1,96 @@
+#load "./build/parameters.cake"
+
+readonly var param = BuildParameters.Instance(Context, "TestContainers");
+
+Setup(context =>
+{
+  var toClean = param.Paths.Directories.ToClean;
+
+  foreach (var project in param.Projects.All)
+  {
+    toClean.Add(project.Path.GetDirectory().Combine("obj"));
+    toClean.Add(project.Path.GetDirectory().Combine("bin"));
+    toClean.Add(project.Path.GetDirectory().Combine("Release"));
+    toClean.Add(project.Path.GetDirectory().Combine("Debug"));
+  }
+
+  Information("Building version {0} of .NET Testcontainers ({1})", param.Version, param.Branch);
+});
+
+Teardown(context =>
+{
+});
+
+Task("Clean")
+  .Does(() =>
+{
+  var deleteDirectorySettings = new DeleteDirectorySettings();
+  deleteDirectorySettings.Recursive = true;
+  deleteDirectorySettings.Force = true;
+
+  foreach (var directory in param.Paths.Directories.ToClean)
+  {
+    if (DirectoryExists(directory))
+    {
+      DeleteDirectory(directory, deleteDirectorySettings);
+    }
+  }
+});
+
+Task("Restore-NuGet-Packages")
+  .Does(() =>
+{
+  DotNetCoreRestore(param.Solution, new DotNetCoreRestoreSettings
+  {
+    Verbosity = param.Verbosity
+  });
+});
+
+Task("Build-Informations")
+  .Does(() =>
+{
+  foreach (var project in param.Projects.All)
+  {
+    Console.WriteLine($"{project.Name} [{project.Path.GetDirectory()}]");
+  }
+});
+
+Task("Build")
+  .Does(() =>
+{
+  DotNetCoreBuild(param.Solution, new DotNetCoreBuildSettings
+  {
+    Configuration = param.Configuration,
+    Verbosity = param.Verbosity,
+    NoRestore = true
+  });
+});
+
+Task("Test")
+  .Does(() =>
+{
+  foreach(var testProject in param.Projects.OnlyTests)
+  {
+    DotNetCoreTest(testProject.Path.FullPath, new DotNetCoreTestSettings
+    {
+      Configuration = param.Configuration,
+      Verbosity = param.Verbosity,
+      NoRestore = true,
+      NoBuild = true,
+      Logger = "trx",
+      ResultsDirectory = param.Paths.Directories.TestResults,
+      ArgumentCustomization = args => args
+        .Append("/p:CollectCoverage=true")
+        .Append("/p:CoverletOutputFormat=opencover")
+        .Append($"/p:CoverletOutput=\"{MakeAbsolute(param.Paths.Directories.TestCoverage)}/\"")
+    });
+  }
+});
+
+Task("Default")
+  .IsDependentOn("Clean")
+  .IsDependentOn("Restore-NuGet-Packages")
+  .IsDependentOn("Build")
+  .IsDependentOn("Test");
+
+RunTarget(param.Target);
