+++ src/TestContainers/Core/Containers/ContainerLaunchException.cs
