+++ src/TestContainers/Core/Containers/RedisContainer.cs
