+++ src/TestContainers/DockerClient/DockerClientProviderStrategy.cs
