+++ src/TestContainers/DockerClient/Helpers/EnvironmentHelper.cs
