+++ src/TestContainers/DockerClient/UnixSocketClientProviderStrategy.cs
