+++ src/TestContainers/DockerClient/WindowsClientProviderStrategy.cs
