+++ test/TestContainers.Tests/ContainerTests/RedisWithConfigTests.cs
