+++ test/TestContainers.Tests/TestContainers.Tests.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>netcoreapp2.2</TargetFrameworks>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="coverlet.msbuild" Version="2.6.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.1.0" />
+    <PackageReference Include="MySql.Data" Version="6.10.6" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
+    <PackageReference Include="xunit" Version="2.4.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="../../src/TestContainers/TestContainers.csproj" />
+  </ItemGroup>
+  <Target Name="CopyRedisConfig" AfterTargets="AfterBuild">
+    <Copy SourceFiles="../RedisConfigs/master-6379.conf" DestinationFolder="$(Outdir)" />
+  </Target>
+  <Import Project="$(MSBuildThisFileDirectory)../../Shared.msbuild" />
+</Project>
